feat: filter by category and tag, one page for photos
